TVGUI-17 Rename Login module
